refactor(publisher): move the site CSS bundle server out of the dispatcher
`server/router.ts` owned `serveSiteCss` along with its `(bundle, hash)`
memo, in-flight de-duplication, and the published-snapshot rebuild walk —
about 130 lines of publishing logic in a module whose one reason is to
dispatch requests. That also left the dispatcher three lines under the
700-line ceiling `module-size-budgets.test.ts` enforces, so any new route
had to displace something first.

The block moves verbatim to `server/publish/siteCssServer.ts`, next to the
`siteCssBundle.ts` that builds what it serves; `tryServeSiteCssNamespace`
now just forwards the path. No behaviour change — same disk-first order,
same memo semantics, same responses.
